It seems there is no nice way to get the monad laws to hold
If we want right_associativity to hold, we need to make sure Bind & Return
return a unique postcond set.
If we use option to encode this, we can't put a (A -> Prop) inside it,
because we can't fold over an (A -> Prop) and arrive in an option, so
we have to put list in it, but then associativity becomes a fold_right
nightmare, and the fold_right will probably pop up elsewhere too, and it's
inconvenient to use fold_right because it "threads" through the set in a
very particular order just to assert no Nones and to take the union,
this seems cumbersome to reason about.
But if we choose not to use option, but set and then ensure uniqueness,
we have to add many uniqueness conditions (not evern sure if possible),
which is cumbersome as well.