Add reservation mechanism to Machine.RiscvMachine
Nothing is added to Spec.Primitives. The instances of
Machine.RiscvMachine in Platform use fail_hard to implement the
reservation functions where appropriate.

The `MetricLog` was not extended with the ability to measure the
reservation operations, because there is currently no interest in doing
so. Accordingly, the Metric* instances of `RiscvMachine` do nothing with
the metrics on reservation operations.