Use COQPATH for coqutil and allow using COQPATH for riscv-coq
We rename src to src/riscv so that adding ./src/ to COQPATH will
correctly pick up riscv.  We also remove `-Q ../coqutil/src coqutil`
from `_CoqProject` so that we can work with an installed version of
`coqutil` (given by passing `EXTERNAL_DEPENDENCIES=1` to `make`).  We
also add a `.dir-locals.el` file to adjust `COQPATH` in emacs/PG to pick
up the copy of coqutil in the parent directory.  Note that this all
requires a very recent version of coqutil.