Experiment: Replace the free monad by the "postcondition monad"
If we use the free monad and have a monadic computation m, an initial
state s, and a postcondition post1, the corresponding verification
condition is `interp m s post1`.
Given this shape of the VC, we can apply weakening to obtain
`interp m s post1` (if interp uses post only in positive positions).

With the "postcondition monad", however, the VC is just `m s post1`,
which is simpler, but comes at the cost that this shape does not
allow us to apply weakening: We'd need a separate judgment keeping
track of the fact that m was constructed using only primitives of
the RiscvProgram typeclass.